requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.6

# Database and data processing
sqlalchemy==2.0.23
pyodbc==4.0.35

# AI and analysis
openai==0.27.0
transformers==4.52.4
torch==2.4.0

# Azure Cognitive Services
azure-cognitiveservices-language-textanalytics==2.0.0
azure-cognitiveservices-language-luis==2.0.0
azure-cognitiveservices-vision-computervision==2.0.0
azure-cognitiveservices-speech==1.31.0
azure-ai-textanalytics==5.3.0
azure-ai-formrecognizer==3.3.0

# Text processing and fuzzy matching
textblob==0.17.1
rapidfuzz==3.6.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
jsonschema==4.17.3
websockets==12.0
fix-busted-json>=0.2.0
